Update scraper.py
I made it where you can tell it to put the author notes in a new file label 

(whatever you named the story file) author notes.txt

If you tell it no, then it will just do as normal but bypass the author notes. 

If the author notes are in the story itself, there isn't much I can do about that.