Update scraper.py
I added the ability for it to use sub folders when writing it. Now it will ask you if you want to add your story to a sub folder. If it already made a sub folder prior, ask if you want to use that name again so you don't have to type it out every time. It saves the name in a new txt file.

I did this to make it more clean and easier to deal with when dealing with multiple stories.